+++ changelog.md
@@ -42,6 +42,7 @@ Codion Change Log
 - EntityEditComponentPanel.EditorComponent.set(ComponentValue) added, automatically links it to the underlying editor value.
 - EntityTablePanel bug fixed, now disables editing for cells based on entities that are neither searchable nor based on small datasets, since a search field can not be automatically instantiated.
 - EntityDialogs.DefaultEditAttributeDialogBuilder refactored, EditAttributePanel added, now handles async properly, validation improved.
+- EditAttributePanel now sets the validation message as component tooltip.
 
 ## 0.18.54
 ### is.codion.common.db